feat: add solution for esm.mjs
